[build fail] reorganization
