+++ src/utils/constants.hpp
@@ -1,0 +1,122 @@
+enum class AutoJFP : int {
+	NotInAutoJFP = 0,
+	JustStarted = 1,
+	JustRestarted = 2,
+	PlayingLevelAtt1 = 3,
+	PlayingLevel = 4,
+};
+
+// current soundtrack
+const int SOUNDTRACK[] = {
+    614361,
+    569208,
+    678606,
+    592012,
+    109650,
+    595394,
+    85046,
+    112527,
+    169589,
+    32207,
+    64842,
+    306254,
+    386900,
+    167229,
+    895761,
+    620959,
+    431573,
+    623104,
+    349494,
+    91656,
+    574484,
+    575665,
+    467267,
+    575341,
+    490203,
+    510853,
+    835420,
+    741034,
+    830628,
+    436814,
+    485978,
+    505813,
+    621166,
+    813526,
+    723722,
+    591740,
+    638150,
+    557651,
+    286138,
+    581444,
+    223469,
+    621144,
+    739358,
+    495840,
+    516709,
+    234565,
+    590533,
+    76743,
+    378965,
+    691180,
+    465909,
+    117474,
+    459080,
+    513917,
+    622111,
+    564162,
+    498828,
+    530913,
+    489111,
+    1126156,
+    383158,
+    648035,
+    32772,
+    656428,
+    469490,
+    579560,
+    472169,
+    675129,
+    586990,
+    561277,
+    114114,
+    347865,
+    718319,
+    718171,
+    469968,
+    771738,
+    666666,
+    829777,
+    739991,
+    905110,
+    709578,
+    219635,
+    764913,
+    163061,
+    333082,
+    659843,
+    323384,
+    278217,
+    645631,
+    258437,
+    98933,
+    465681,
+    31214,
+    99641,
+    478283,
+    36355,
+    797578,
+    614354,
+    28549,
+    353593,
+    141262,
+    343707,
+    778652,
+    393342,
+    137492,
+    685686,
+    616845,
+    63082,
+    233860,
+    598349,
+    437224
+};
